Oops, reference to double left in common code.

+++ micro/src/org/cougaar/microedition/asset/ControllerResource.java
@@ -41,7 +41,7 @@ public abstract class ControllerResource extends ResourceAdapter
     return units;
   }
 
-  public double getScalingFactor() {
+  public long getScalingFactor() {
     return scalingFactor;
   }
